moved all login and logout files into a folder
